# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/taxonomies": {
      "get": {
        "tags": [
          "Taxonomy"
        ],
        "summary": "To create a new taxonomy",
        "operationId": "getTaxonomies",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "201": {
            "description": "Request created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyInfo"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Taxonomy"
        ],
        "summary": "To create a new taxonomy",
        "operationId": "createTaxonomyInfo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxonomyInfo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Request created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyInfo"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "409": {
            "description": "Conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Taxonomy"
        ],
        "summary": "To delete a taxonomy from file",
        "operationId": "deleteTaxonomy",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "200": {
            "description": "Request deleted.",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        }
      }
    },
    "/taxonomies/upload": {
      "post": {
        "tags": [
          "Taxonomy"
        ],
        "summary": "To create a new taxonomy from file",
        "operationId": "createFromFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyInfo"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "409": {
            "description": "Conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TaxonomyInfo": {
        "required": [
          "CODICE TIPO ENTE CREDITORE",
          "CODICE TIPOLOGIA SERVIZIO",
          "DATA FINE VALIDITA",
          "DATA INIZIO VALIDITA",
          "DATI SPECIFICI DI INCASSO",
          "DESCRIZIONE MACRO AREA",
          "DESCRIZIONE TIPO SERVIZIO",
          "Motivo Giuridico della riscossione",
          "NOME MACRO AREA",
          "Progressivo Macro Area per Ente Creditore",
          "TIPO ENTE CREDITORE",
          "TIPO SERVIZIO",
          "VERSIONE TASSONOMIA"
        ],
        "type": "object",
        "properties": {
          "CODICE TIPO ENTE CREDITORE": {
            "pattern": "\\d+",
            "type": "string",
            "description": "CODICE TIPO ENTE CREDITORE",
            "format": "integer",
            "example": "01"
          },
          "TIPO ENTE CREDITORE": {
            "type": "string",
            "description": "TIPO ENTE CREDITORE"
          },
          "Progressivo Macro Area per Ente Creditore": {
            "pattern": "\\d+",
            "type": "string",
            "description": "Progressivo Macro Area per Ente Creditore",
            "format": "integer",
            "example": "01"
          },
          "NOME MACRO AREA": {
            "type": "string",
            "description": "NOME MACRO AREA"
          },
          "DESCRIZIONE MACRO AREA": {
            "type": "string",
            "description": "DESCRIZIONE MACRO AREA"
          },
          "CODICE TIPOLOGIA SERVIZIO": {
            "pattern": "\\d+",
            "type": "string",
            "description": "CODICE TIPOLOGIA SERVIZIO",
            "format": "integer",
            "example": "100"
          },
          "TIPO SERVIZIO": {
            "type": "string",
            "description": "TIPO SERVIZIO"
          },
          "Motivo Giuridico della riscossione": {
            "type": "string",
            "description": "Motivo Giuridico della riscossione"
          },
          "DESCRIZIONE TIPO SERVIZIO": {
            "type": "string",
            "description": "DESCRIZIONE TIPO SERVIZIO"
          },
          "VERSIONE TASSONOMIA": {
            "pattern": "\\d+",
            "type": "string",
            "description": "VERSIONE TASSONOMIA",
            "format": "integer",
            "example": "1"
          },
          "DATI SPECIFICI DI INCASSO": {
            "type": "string",
            "description": "DATI SPECIFICI DI INCASSO"
          },
          "DATA INIZIO VALIDITA": {
            "pattern": "dd/MM/yyyy",
            "type": "string",
            "description": "DATA INIZIO VALIDITA",
            "example": "25/07/2022"
          },
          "DATA FINE VALIDITA": {
            "pattern": "dd/MM/yyyy",
            "type": "string",
            "description": "DATA FINE VALIDITA",
            "example": "25/07/2022"
          }
        }
      },
      "ProblemJson": {
        "required": [
          "detail",
          "status",
          "title"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "format": "int32",
            "example": 200
          },
          "title": {
            "type": "string",
            "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
          },
          "detail": {
            "type": "string",
            "description": "A human readable explanation specific to this occurrence of the problem.",
            "example": "There was an error processing the request"
          }
        }
      }
    }
  }
}