# src/main/kotlin/it/pagopa/gov/taxonomy/model/ProblemJson.kt
package it.pagopa.gov.taxonomy.model

import com.fasterxml.jackson.annotation.JsonIgnoreProperties
import com.fasterxml.jackson.annotation.JsonInclude
import com.fasterxml.jackson.annotation.JsonProperty
import io.swagger.v3.oas.annotations.media.Schema
import lombok.*
import javax.validation.constraints.Max
import javax.validation.constraints.Min

/**
 * Object returned as response in case of an error.
 *
 *  See [it.pagopa.pagopa.apiconfig.exception.ErrorHandler]
 */
@ToString
@JsonIgnoreProperties(ignoreUnknown = true)
data class ProblemJson(
    @JsonProperty("status")
    @Schema(
        example = "200",
        description = "The HTTP status code generated by the origin server for this occurrence of the problem."
    )
    val status: @Min(100) @Max(600) Int,

    @JsonProperty("title")
    @Schema(description = "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable")
    val title: String,

    @JsonProperty("detail")
    @Schema(
        example = "There was an error processing the request",
        description = "A human readable explanation specific to this occurrence of the problem."
    )
    val detail: String
)